pyproject.toml
[project]
name = "documind"
version = "0.1.0"
description = "OCR-first document intelligence platform"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.88.0,<1.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "google-genai>=1.70.0,<2.0.0",
    "httpx>=0.28.1,<1.0.0",
    "numpy>=1.26.0,<3.0.0",
    "openai>=2.30.0,<3.0.0",
    "rank-bm25>=0.2.2,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"
